pgcluu CGI will now append new csv statistic to cached binary files. When building reports, pgcluu will first load statistics from binary files and then complete statistics with data from csv file collected after the last mode cache run. This mean that if cache mode is run each minute on an incremental statistics collect (option -r or -R of pgcluu_collectd), pgcluu is able to create report very quickly at any time.
The menu is now built at end to avoid reading csv files first to look
for database, disk device and network interface.

The home page is now just built from the binary files to speed up the
first screen. As we do not append last data collected from the csv files
we indicate at which time the cache was last built. This is to give a
quick snapshot of the dashboard. When looking at the statistics binary
and csv files are both read. If the cache have not been run yet, pgcluu
will read dashboard statistics from the csv file which can takes much
longer to display.