Dockerise pgcluu
pgcluu container helps run both the collection and the report generator.

Simply run it and supply a volume for the report, with your db details

  # docker run --rm -it \
	-v /dockerhost/report:/tmp/pgcluu_report:z,rw \
	-e POSTGRES_HOSTNAME=hostname_to_analyze \
	-e POSTGRES_DB=dbname \
	 pgcluu

This will start pgcluu_collectd on the desired db in the container
forground untill you stop it with CTRL-\ which will start generating the
report into the PGCLUU_REPORT_DIR and then the container will terminate.

Full list of environment variables for customization:
	PGCLUU_STATS_DIR
	PGCLUU_REPORT_DIR
	POSTGRES_USERNAME
	POSTGRES_PASSWORD
	POSTGRES_DB
	POSTGRES_HOSTNAME
	POSTGRES_PORT

fixes: #78
Signed-off-by: Roy Golan <rgolan@redhat.com>