Add storage information about PostgreSQL packages installed on the system.
Add command line option --package-list to be able to set a custom command
to list PostgreSQL packages. Default is to autodetect package type between
rpm and dpkg, using command 'rpm -qa' or 'dpkg -l'. If you have an other
system you can use this option to set a custom command. A filter on keyword
'postgres' is appended to the command: ' | grep postgres'.