Add a comment in the CREATE INDEX when the suggested index may be redundant with existing ones.
Suggested missing indexes on FK may be redundant with other ones
(typically: an index with more columns already exists).
Add a comment in the DDL pointing to these indexes,
so the DBA can decide itself if the new index is worth it or not.
There is no check with other suggested indexes that do not exist yet.