[pgcluu_collectd] dump pg_stats view
It is useful to dump the pg_stats views. The dump can be imported
in a table created like this:

CREATE TABLE my_client_pgstats (
  export_time TEXT,
  dbname TEXT,
  schemaname TEXT,
  tablename TEXT,
  attname TEXT,
  inherited BOOLEAN,
  null_frac REAL,
  avg_width INT,
  n_distinct REAL,
  most_common_vals TEXT[],
  most_common_freqs REAL[],
  histogram_bounds TEXT[],
  correlation REAL
);

Note that the columns most_common_vals and histogram_bounds have
the type TEXT[] instead of anyarray.

The point is for the DBA to study the statistics, and nothing else.